Refactor `ATOF`
Return positive zero if a string representing negative zero is passed to `ATOF`.
This is because we don't support all features of IEEE 754, including negative zero.
As a result, if negative zero is passed to any arithmetic subroutine, it won't be recognized as zero and will unpack into a non-zero value.